+++ src/coding/coding.h
@@ -154,19 +154,45 @@ void decode_mp4_aac(mp4_aac_codec_data * data, sample * outbuf, int32_t samples_
 void decode_at3plus(VGMSTREAM *vgmstream, sample * outbuf, int channelspacing, int32_t samples_to_do, int channel);
 #endif
 
-/* ffmpeg_decoder */
+
 #ifdef VGM_USE_FFMPEG
+/* ffmpeg_decoder */
 void decode_ffmpeg(VGMSTREAM *stream, sample * outbuf, int32_t samples_to_do, int channels);
 void reset_ffmpeg(VGMSTREAM *vgmstream);
 void seek_ffmpeg(VGMSTREAM *vgmstream, int32_t num_sample);
 
 void ffmpeg_set_skip_samples(ffmpeg_codec_data * data, int skip_samples);
 
+/* ffmpeg_decoder_utils */
+int ffmpeg_fmt_chunk_swap_endian(uint8_t * chunk, uint16_t codec);
 int ffmpeg_make_riff_atrac3(uint8_t * buf, size_t buf_size, size_t sample_count, size_t data_size, int channels, int sample_rate, int block_align, int joint_stereo, int encoder_delay);
 int ffmpeg_make_riff_xma1(uint8_t * buf, size_t buf_size, size_t sample_count, size_t data_size, int channels, int sample_rate, int stream_mode);
 int ffmpeg_make_riff_xma2(uint8_t * buf, size_t buf_size, size_t sample_count, size_t data_size, int channels, int sample_rate, int block_count, int block_size);
 int ffmpeg_make_riff_xma2_from_fmt(uint8_t * buf, size_t buf_size, off_t fmt_offset, size_t fmt_size, size_t data_size, STREAMFILE *streamFile, int big_endian);
 int ffmpeg_make_riff_xwma(uint8_t * buf, size_t buf_size, int codec, size_t sample_count, size_t data_size, int channels, int sample_rate, int avg_bps, int block_align);
+
+/* XMA sample parser info (struct to avoid passing so much stuff, separate for reusing) */
+typedef struct {
+    int xma_version;
+    int channels;
+    int stream_mode;
+    off_t data_offset;
+    size_t data_size;
+    int loop_flag;
+    /* frame offsets */
+    uint32_t loop_start_b;
+    uint32_t loop_end_b;
+    uint32_t loop_start_subframe;
+    uint32_t loop_end_subframe;
+
+    /* output */
+    int32_t num_samples;
+    int32_t skip_samples;
+    int32_t loop_start_sample;
+    int32_t loop_end_sample;
+} xma_sample_data;
+void xma_get_samples(xma_sample_data * xma, STREAMFILE *streamFile);
+
 #endif
 
 #endif /*_CODING_H*/
